Getting all movies directed by an input director
I couldn't find a better way to do it, I've tried to find the id of the director with web scrapping but I was not capable to, I also tried to find all movies directed by the inputted director trough the API, but I was also not able to. I'm gonna try web scrap everything tho in the future.